test mysql binary to new file
