+++ .github/workflows/mindsdb.yml
@@ -72,6 +72,10 @@ jobs:
           # MySQL API
           echo -e "\n===============\ntest MySQL API\n===============\n"
           python tests/integration_tests/flows/test_mysql_api.py
+        
+          # MySQL binary API
+          echo -e "\n===============\ntest MySQL binary API\n===============\n"
+          python tests/integration_tests/flows/test_mysql_bin_api.py
 
           # # Kafka Stream
           # echo -e "\n===============\ntest Kafka Stream\n===============\n"
